simplified the file structure
