+++ lib/src/lib/server/handlers.py
@@ -17,15 +17,7 @@ Creates an Api Endpoint that allows communication with the Storage service.
 """
 import grpc
 import sys
-
-from storage.server.interfaces import Service, ServiceFactory
-
-# Handlers
-from storage.protos import storage_pb2_grpc
-
-# Services to register
-from storage.server.services import storage
-
+from lib.stubs.interfaces import Stub
 
 def lazy_get_handler(service: str):
     """Returns the handler of the service, generated by the .proto file"""
@@ -36,7 +28,7 @@ def lazy_get_handler(service: str):
 
 
 def _add_service_endpoints_to_server(
-    name: str, service: Service, server: grpc.Server
+    name: str, stub: Stub, server: grpc.Server
 ) -> None:
     """Retrieves the API endpoint for some class and adds it to the server"""
 
@@ -47,14 +39,13 @@ def _add_service_endpoints_to_server(
     add_handler = getattr(handler, f"add_{name}Servicer_to_server")
 
     # Add the handler to the server
-    add_handler(service(), server)
+    add_handler(stub(), server)
 
 
-def add_handlers(server: grpc.Server) -> grpc.Server:
+def add_handlers(server: grpc.Server, stubs: dict[str, Stub]) -> grpc.Server:
     """Add the service handlers to a server"""
-    services: dict = ServiceFactory.services
 
-    for name, service in services.items():
-        _add_service_endpoints_to_server(name=name, service=service, server=server)
+    for name, stub in stubs.items():
+        _add_service_endpoints_to_server(name=name, stub=stub, server=server)
 
     return server
