+++ workspaces/scraper/src/scraper/scraper/scraper.py
@@ -12,179 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import inspect
-import re
-from abc import abstractmethod
-from dataclasses import dataclass, field
-import types
-from typing import Any, Protocol, Sequence
 
-from bs4 import BeautifulSoup
+from typing import Any
+
+from lib.scraper.scraper import Scraper
+from lib.scraper.factory import ScraperFactory
 from scraper.scraper.blueprint import Blueprint, get_blueprint
-
-
-@dataclass
-class ScraperProtocol(Protocol):
-    @abstractmethod
-    def scrape(self, structure: Sequence[dict[Any, Any]]) -> dict[Any, Any]:
-        """Logic of the scrapping process"""
-        raise NotImplementedError
-
-    @abstractmethod
-    def map_properties(self, prop) -> dict[Any, Any]:
-        """Map the properties to include regular expressions"""
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_attribute(self, content, attribute):
-        """Get some attribute from content"""
-        raise NotImplementedError
-
-    @abstractmethod
-    def process(
-        self,
-        content,
-        instructions: Sequence[dict[Any, Any]],
-        clean_expr: str = None,
-    ):
-        """Wrapper for processing instructions"""
-        raise NotImplementedError
-
-    @abstractmethod
-    def clean(self, data: str, expression: str = None) -> str:
-        """Method to clean the results utilising a regular expression"""
-        raise NotImplementedError
-
-
-@dataclass
-class Scraper(ScraperProtocol):
-    """Base Scraper class that implements the scraper protocol
-
-    It can not be used alone, as it does not implement the "scrape" method
-    """
-
-    content: BeautifulSoup = field(default_factory=BeautifulSoup)
-
-    @classmethod
-    def from_html(cls, html: str):
-        content = BeautifulSoup(html, "lxml")
-        return cls(content=content)
-
-    def clean(self, data: str, expression: str = None) -> str:
-        if expression:
-            exp = re.compile(str(expression))
-            groups = exp.search(data)
-            if groups:
-                ret = groups.group()
-                data = ret
-
-        return data
-
-    def process(
-        self,
-        content,
-        instructions: Sequence[dict[Any, Any]],
-        clean_expr: str = None,
-    ):
-        # Continue only if there is content or an instruction
-        iscontent = hasattr(content, "find_all")
-        if not (iscontent and instructions):
-            return content
-
-        instruction = instructions.pop(0)
-        results: Sequence = []
-
-        # Find the element(s)
-        props = instruction.get("props")
-        props = self.map_properties(props)
-        found = content.find_all(**props)
-
-        # Get the attribute(s)
-        attrs = instruction.get("attrs", [])
-
-        for el in found:
-            for attr in attrs:
-                el = self.get_attribute(content=el, attribute=attr)
-
-            # Copy the list of instructions for each of the results
-            # NOTE: It will cause the results to create branches and groups
-            # of results
-            cp = instructions.copy()
-            branch = self.process(el, cp, clean_expr)
-
-            # If the branch is a single value, add it to the list
-            # of results, otherwise, join the list of the results
-            if isinstance(branch, str):
-                res = self.clean(branch, clean_expr)
-                results.append(res)
-            else:
-                if branch:
-                    results += branch
-
-        return results
-
-    def map_properties(self, prop) -> dict[Any, Any]:
-
-        if hasattr(prop, "items"):
-            # Iterate through the property items
-            for k, v in prop.items():
-
-                # If the property is an object, return this method
-                if isinstance(v, (dict, object)):
-                    prop[k] = self.map_properties(v)
-
-        else:
-            if isinstance(prop, str):
-                prop = re.compile(str(prop))
-
-        return prop
-
-    def get_attribute(self, content, attribute):
-        # Get the attribute from the object or access the property
-        attr = getattr(content, attribute, None)
-
-        # If the attribute was a method rather than an attribute, call it.
-        if inspect.ismethod(getattr(content, attribute)):
-            attr = attr()
-
-        if not attr:
-            attr = content.get(attribute)
-
-        # Remove extra spaces at both ends of the string
-        if isinstance(attr, str):
-            attr = attr.strip()
-
-        return attr
-
-    def hasmethod(self, obj, name):
-        return hasattr(obj, name) and type(getattr(obj, name)) == types.MethodType
-
-
-class ScraperFactory:
-    scrapers = {}
-
-    @classmethod
-    def get_scraper(cls, scraper: str) -> Scraper:
-        try:
-            retval = cls.scrapers[scraper]
-        except KeyError as err:
-            raise NotImplementedError(f"{scraper=} doesn't exist") from err
-        return retval
-
-    @classmethod
-    def register(cls, name):
-        def decorator(decorator_cls: Scraper):
-            cls.scrapers[name] = decorator_cls
-            return decorator_cls
-
-        return decorator
-
 
 @ScraperFactory.register("simple")
 class SimpleScraper(Scraper):
     """Scraper for 'simple' tagged structures"""
 
-    def scrape(self, structure: Sequence[dict[Any, Any]]) -> dict[Any, Any]:
+    def scrape(self, structure: list[dict[Any, Any]]) -> dict[Any, Any]:
         # Initialise a variable to hold the data points
         data: dict = {}
 
@@ -213,7 +52,7 @@ class SimpleScraper(Scraper):
 class GroupedScraper(Scraper):
     """Scraper for content that contains groups of fields"""
 
-    def scrape(self, structure: Sequence[dict[Any, Any]]) -> dict[Any, Any]:
+    def scrape(self, structure: list[dict[Any, Any]]) -> dict[Any, Any]:
         data: dict = {}
 
         for group in structure:
@@ -275,5 +114,4 @@ def scrape(
 
     # scrape the content
     content: dict[Any, Any] = scraper.scrape(structure=blueprint.structure["struct"])
-
     return content
