--- workspaces/storage/src/storage/client/interfaces.py
@@ -1,19 +1,0 @@
-# Copyright 2023 Ricardo Yaben
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Protocol
-
-class Scraper(Protocol):
-    def scrape(self, model: str, market: str, data: bytes):
-        raise NotImplementedError
